backend/requirements.txt
backend==0.2.4.1
Flask==3.1.0
flask_cors==5.0.1
numpy==2.2.3
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Requests==2.32.3
torch==2.6.0
gunicorn==23.0.0
tqdm==4.67.1